zoom added

+++ src/data_types.rs
@@ -1,4 +1,4 @@
 // rgba
 pub type Colour = u32;
-pub type Vec2 = (i32, i32);
+pub type Vec2 = (f32, f32);
 pub type LogicalNodeHandle = usize;
